pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "py-pilecore"
version = "0.0.1"
description = "Public python SDK for the CEMS PileCore web-API."
requires-python = ">=3.9"
dependencies = [
    'pygef>0.8,<1',
    'numpy>1,<2',
    'pandas>2,<3',
    'cems-nuclei[client]>0.3,<1',
    'matplotlib>3,<4',
]
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["API", "PileCore", "CEMS", "CRUX"]

[project.urls]
repository = "https://github.com/cemsbv/py-pilecore"

[project.optional-dependencies]
test = ["coveralls", "pytest"]
lint = ["black", "isort", "mypy", "flake8"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
files = ["pypilecore"]
mypy_path = 'src'
namespace_packages = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true
disallow_subclassing_any = true
no_implicit_optional = true
disallow_untyped_defs = true
strict_concatenate = true
disallow_any_generics = false
disallow_untyped_calls = true
warn_return_any = false
no_implicit_reexport = true
strict_equality = true
